Removed zsh files.
I am no longer using zsh.
I only ever used it because I couldn't be bothered to configure my macOS
to use bash. I really want to learn one shell at a time.

As an aside, I've learned disturbing things about how zsh does
"word-splitting": https://unix.stackexchange.com/questions/26661/what-is-word-splitting-why-is-it-important-in-shell-programming
I find this disturbing mostly because I think the idea of word splitting
and the other shell features that lead to it are flawed beyond repair,
and at this moment in this night, it seems that zsh's change is something
I'd be likely to forget, would be *painful* to discover, and may not
significantly increase the usability of anything.