amend! amend! Instructions for how to handle README.md in the root.
Instructions for how to handle README.md in the root.

If you don't want it in your home directory, that is. Which I don't!
It's real neat git has this feature.

amendment: Fixed a bunch of silly mistakes with these instructions.

amendment: Fixed how README.md refers to
`~/.samuel_skean_dotfiles_miscellanea`.