Added LICENSE.txt, documented it (including here).
I was looking for a simple, short, permissive license that lets people do what they want with my code.
I looked at https://www.gnu.org/licenses/license-list.en.html.
I tried to keep an open mind to their recommendations despite not liking their attitude as an organization.
Most of their comments on licenses that met my basic criteria recommended the Apache License 2.0. I found that license too long and wordy.
I strongly considered the 2-clause BSD license and the 3-clause BSD license, but I found the argument against their potentially confusing name somewhat compelling.

\[ UNFORTUNATE FLAW: The next paragraph is a bit long and redundant. I don't feel like fixing this. \]

I saw many recommendations elsewhere of the MIT license (e.g. on https://choosealicense.com/ and [on opensource.guide](https://opensource.guide/legal/#which-open-source-license-is-appropriate-for-my-project), but I disliked its wordy enumeration of all the things someone could do with the software, and I was *again* swayed by the FSF's argument that its name might cause confusion. I much prefer the way the BSD licenses simply say "use". I had seen [this](https://opensource.stackexchange.com/questions/217/what-are-the-essential-differences-between-the-bsd-and-mit-licences) stackexchange post before any of that, which argues that "use" seems legally equivalent (I, woefully uninformed, agree), and the FSF, which seems to pay plenty of lawyers, made plenty of noise to the effect that the BSD licenses are largely equivalent to the MIT license in the freedoms they bestow on the recipient of the source.
I found this license (the Boost Software License 1.0). I like that its name appears unambiguous. I like that it's mentioned on [this Github page that lists licenses](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/licensing-a-repository) and [on opensource.org](https://opensource.org/license/bsl-1-0). I don't love that it enumerates so many different things a recipient of the source code can do - I still find that unnecessarily wordy and potentially confusing. At the same time, I can see a bit of appeal to it - it does seem more "professional", somehow, and potentially clearer to some people - but that's mostly me trying to reduce cognitive dissonance. I *really* like that it doesn't require the license to be present on "machine-executable object code" - even though that probably doesn't really apply to this repo. I just don't want to burden anyone with putting a license next to any executable.

Plus, I have a fondness for the license's "steward" (as per [opensource.guide's listing](https://opensource.guide/legal/#which-open-source-license-is-appropriate-for-my-project)) - the Boost C++ libraries (www.boost.org). My friends have written cool software with them, and they're a long-standing project trying to marry expressive general-purpose APIs with fast performance and low-level control.
In choosing this license, I was really thinking of choosing a license for all software I copyright, to make my future life easier (and potentially, to make the lives of any users of that software easier).