snake_case -> kebab-case for filenames by default
Also my name is `samuel-skean` in filenames and elsewhere [that's cause
it is on github, and can be on every machine I control :)].

And "by default" means "where I control the filename" (so, not
`bash_profile`) and where there's no other convention/prevention (the
convention for this `lua` *code* is to use camelCase (maybe that's 'wrong').
While it sure seems like `lua` - at *least* for `nvim` configuration - can
use any valid filenames to store its code, the filenames that correspond
to elements in the code that cannot be named with kebab-case should be
written as those entities in the code are written - here, that means
camelCase.
The name `samuel-skean` as a directory in my `nvim` config code is not
just an override of the current standard for that code to be camelCase
so I can be vain about the consistency of my name,
it's justified by not directly corresponding to an entity in the code
(besides strings that name that directory).

And all mentions of those filenames should (!!) be updated too within
the contents of any files.